move out the autoconfigure functions
